requirements.txt
numpy==2.1.3
pymdptoolbox==4.0b3
scipy==1.14.1
